(Feature): Camera follows player. (Fix): FOV Mask Follow
Added logic so the camera follows the player around. Fixed logic so the FOV mask follows the player. (BUGGED) after starting the game the objects become invisible in the scene (they are considered visible by debbuging yet they dont appear on screen, the camera and mask follow the player, and the player can interact with the objectives. Probably a layer order or mask bug).